% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotKernels}
\alias{plotKernels}
\title{Just a display of what the different kernels look like.}
\usage{
plotKernels(windowSize = 50)
}
\arguments{
\item{windowSize}{window size on the two sides of the center.}
}
\description{
Just a display of what the different kernels look like.
}

