% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dipPeaks}
\alias{dipPeaks}
\title{Estimate density and find peaks}
\usage{
dipPeaks(bamFile, bigWigOut = NULL, chromInfo = NULL,
  scratchDirBase = "~", outDir = "~", cores = 1, perChromCutoff = FALSE,
  windowSize = 50, windowStep = 5, indexFile = NULL, retainTemp = FALSE,
  limitChrom = NULL)
}
\arguments{
\item{bamFile}{Input file in bam format}

\item{bigWigOut}{Produce a smoothed output density file in bigwig format (using wigToBigWig)?}

\item{indexFile}{Index file made from "samtools index" on the bam input file; deaults to bamFile.bai, or you can provide one.}

\item{scratchDirBase}{Directory for scratch output; use local disk for optimal processing. dipPeak writes a significant amount of data to disk as it runs, to reduce memory use. This I/O takes time, and will be faster if you use a local disk or fast mount spot.}

\item{outDir}{Final folder (the files to keep will be moved here from the scratch dir).}

\item{cores}{Add additional cores (up to 23) for faster processing. (Default:1)}

\item{useRle}{Boolean; offers speed and memory improvement}

\item{perChromCutoff}{Boolean; If yes, calculates a per-chromosome; otherwise, defaults to genome-wide cutoff.}

\item{windowSize}{Choose the size of the smoothing window (default:50)}

\item{windowStep}{Choose the window step (5 will take a window every 5 bases) (default:5)}

\item{chromInfo}{UCSC chromInfo.txt file (required only for bigWigOut)}

\item{retainTemp}{Do you want to retain temporary files?}

\item{limitChrom}{You can limit the smoothing to a list of chromosomes (defaults to everything in the bam file)}
}
\description{
Calculates a smoothed track from input count data,
and identifies peaks in the data. Outputs the peaks and smoothed
track to file.
}
\examples{
dipPeaks("sequences.bam", bigWigOut="density.bw", outDir="outDir", cores=8);
}

